# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для генерации персонализированных поздравлений: AI-изображения + песни. Аутентификация через httpOnly cookie (access_token).",
        "title": "FunBaza API",
        "contact": {
            "name": "FunBaza Team"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/generations": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Возвращает постраничный список всех генераций текущего пользователя.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "generations"
                ],
                "summary": "Список генераций",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Лимит (макс. 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Смещение",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.GenerationListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Запускает задачу генерации изображений и/или песни. Списывает кредиты. Возвращает ID задачи для последующего опроса статуса через GET /generations/:id/status.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "generations"
                ],
                "summary": "Создать генерацию",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID существующей сессии (если нет — создаётся новая)",
                        "name": "session_id",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "UUID родительской генерации в треде",
                        "name": "parent_id",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "default": 3,
                        "description": "Количество изображений (0-3)",
                        "name": "image_count",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Количество песен (0-3)",
                        "name": "song_count",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Текстовое описание для генерации изображений",
                        "name": "image_prompt",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "default": "gpt-image-2",
                        "description": "Модель генерации изображений: gpt-image-2, flux-2-flex, seedream-5-lite",
                        "name": "image_model",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Промт для генерации текста песни (если song_lyrics не задан)",
                        "name": "song_prompt",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Текст песни (если задан — song_prompt игнорируется)",
                        "name": "song_lyrics",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Стиль песни (pop, jazz, rock и т.д.)",
                        "name": "song_style",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Фото пользователя JPG/PNG до 10MB (макс. 3)",
                        "name": "photos[]",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateGenerationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "402": {
                        "description": "Недостаточно кредитов",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/generations/lyrics": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Генерирует текст и заголовок песни по промту через kie.ai. Требует KIE_API_KEY.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "generations"
                ],
                "summary": "Сгенерировать текст песни",
                "parameters": [
                    {
                        "description": "Промт",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.LyricsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.LyricsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Генератор текста недоступен",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.CreateGenerationResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "session_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440001"
                },
                "status": {
                    "type": "string",
                    "example": "pending"
                }
            }
        },
...
        "handlers.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "unauthorized"
                },
                "message": {
                    "type": "string",
                    "example": "Refresh token missing"
                }
            }
        },
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/handlers.ErrorDetail"
                }
            }
        },
...
        "handlers.GenerationListResponse": {
            "type": "object",
            "properties": {
                "generations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.GenerationRequest"
                    }
                },
                "limit": {
                    "type": "integer",
                    "example": 20
                },
                "offset": {
                    "type": "integer",
                    "example": 0
                }
            }
        },
...
        "handlers.LyricsRequest": {
            "type": "object",
            "required": [
                "prompt"
            ],
            "properties": {
                "prompt": {
                    "type": "string"
                }
            }
        },
        "handlers.LyricsResponse": {
            "type": "object",
            "properties": {
                "text": {
                    "type": "string",
                    "example": "Куплет 1...\nПрипев..."
                },
                "title": {
                    "type": "string",
                    "example": "Поздравление"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CookieAuth": {
            "type": "apiKey",
            "name": "access_token",
            "in": "cookie"
        }
    }
}